feat: 1.0.0 release - Hin2n 联机, JRE 选择, UI 对齐 HMCL 风格

+++ HMCLPE/src/main/java/com/tungsten/hmclpe/launcher/game/World.java
@@ -113,7 +113,7 @@ public class World {
             throw new IOException("level.dat missing LevelName");
 
         if (data.get("LastPlayed") instanceof LongTag)
-            lastPlayed = data.<LongTag>get("LastPlayed").getValue();
+            lastPlayed = ((Long) data.<LongTag>get("LastPlayed").getValue()).longValue();
         else
             throw new IOException("level.dat missing LastPlayed");
